match: reconstruct the common and map resource loaders
common_resources_load and map_resources_load are exact. Both walk the
chunked COM.DAT / MIXA.DAT / MIXB.DAT streams with an assigning
STREAM_NEXT macro used inside the call arguments, which is the only
spelling that adds the chunk length before the base and schedules the
advance into the delay slot the way retail does. The floor parameter of
map_resource_path_set_floor is an int, not a byte: the caller passes it
unmasked. The 3264-byte common data block is typed as a byte array so its
copy keeps the runtime alignment test, while the level growth table copy
casts to the word-aligned record type.

MIX.TIM becomes a .data claim (map_mix_tim_filename), the five map grid
address sites are admitted to the relocation census, and the two bodies'
other relocation rows are promoted after review.

+++ src/game/map_resource_path_set_floor.c
@@ -4,7 +4,7 @@
 extern char map_resource_path[12];
 
 ADDRESS(0x8001b390, 0x14)
-void map_resource_path_set_floor(u8 floor)
+void map_resource_path_set_floor(s32 floor)
 {
     map_resource_path[1] = floor + '0';
 }
